+++ packages/framework/src/Model/Order/Item/OrderProductFacade.php
@@ -96,7 +96,11 @@ class OrderProductFacade
     public function addOrderProductsToStock(array $orderProducts)
     {
         if ($this->moduleFacade->isEnabled(ModuleList::PRODUCT_STOCK_CALCULATIONS)) {
-            $this->orderProductService->returnOrderProductsToStock($orderProducts);
+            $orderProductsUsingStock = $this->orderProductService->getOrderProductsUsingStockFromOrderProducts($orderProducts);
+            foreach ($orderProductsUsingStock as $orderProductUsingStock) {
+                $product = $orderProductUsingStock->getProduct();
+                $product->addStockQuantity($orderProductUsingStock->getQuantity());
+            }
             $this->em->flush();
             $this->runRecalculationsAfterStockQuantityChange($orderProducts);
         }
