+++ packages/framework/src/Model/Order/Item/OrderProductService.php
@@ -6,18 +6,6 @@ class OrderProductService
 {
     /**
      * @param \Shopsys\FrameworkBundle\Model\Order\Item\OrderProduct[] $orderProducts
-     */
-    public function returnOrderProductsToStock(array $orderProducts)
-    {
-        $orderProductsUsingStock = $this->getOrderProductsUsingStockFromOrderProducts($orderProducts);
-        foreach ($orderProductsUsingStock as $orderProductUsingStock) {
-            $product = $orderProductUsingStock->getProduct();
-            $product->addStockQuantity($orderProductUsingStock->getQuantity());
-        }
-    }
-
-    /**
-     * @param \Shopsys\FrameworkBundle\Model\Order\Item\OrderProduct[] $orderProducts
      * @return \Shopsys\FrameworkBundle\Model\Product\Product[]
      */
     public function getProductsUsingStockFromOrderProducts(array $orderProducts)
